requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.2.0
beautifulsoup4==4.10.0
blinker==1.4
branca==0.4.2
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
Django==4.0
django-jsonfield==1.4.1
django-postgres-extensions==0.9.3
folium==0.12.1.post1
frozenlist==1.2.0
geographiclib==1.52
geopy==2.2.0
gprof2dot==2021.2.21
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
Jinja2==3.0.3
kaitaistruct==0.9
llvmlite==0.37.0
lxml==4.7.1
MarkupSafe==2.0.1
multidict==5.2.0
numba==0.54.1
numpy==1.20.3
outcome==1.1.0
Pillow==8.4.0
progress==1.6
psycopg2-binary==2.9.2
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
PySocks==1.7.1
requests==2.26.0
selenium==4.1.0
selenium-wire==4.5.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
yarl==1.7.2
zstandard==0.16.0
